feat(core): integrate preparator entity and adapter interface types
- Extract StoryObject and StoryObjectMap types into dedicated entity file
- Add isClaudeAvailable method to ClaudeRepository interface
- Add getAllOpened and getStoryObjectMap methods to IssueRepository interface
- Add prepareStatus method to ProjectRepository interface
- Add CopilotRepository adapter interface
- Update all imports to use centralized entity definitions
- Add stub implementations for new interface methods in adapters
