+++ src/domain/usecases/adapter-interfaces/ClaudeRepository.ts
@@ -2,4 +2,5 @@ import { ClaudeWindowUsage } from '../../entities/ClaudeWindowUsage';
 
 export interface ClaudeRepository {
   getUsage(): Promise<ClaudeWindowUsage[]>;
+  isClaudeAvailable(threshold: number): Promise<boolean>;
 }
