+++ src/domain/usecases/adapter-interfaces/CopilotRepository.ts
@@ -1,0 +1,3 @@
+export interface CopilotRepository {
+  run(prompt: string, model: 'gpt-5-mini', processTitle: string): void;
+}
