+++ src/domain/usecases/adapter-interfaces/ProjectRepository.ts
@@ -8,4 +8,5 @@ export interface ProjectRepository {
     storyOption: (Omit<FieldOption, 'id'> & { id: FieldOption['id'] | null })[],
   ) => Promise<FieldOption[]>;
   getByUrl: (url: string) => Promise<Project>;
+  prepareStatus: (name: string, project: Project) => Promise<Project>;
 }
